Fix float action truncation in replay batch conversion

+++ services/learner-py/learner/replay.py
@@ -169,7 +169,7 @@ def sample_response_to_batch(response: SampleResponseLike) -> TransitionBatch:
     # Validate and stack tensor fields with improved error handling
     try:
         obs_tensor = _stack_tensors(observations, field="observation").to(device="cpu")
-        action_tensor = _stack_tensors(actions, field="action").to(device="cpu", dtype=torch.long)
+        action_tensor = _stack_tensors(actions, field="action").to(device="cpu")
     except ValueError as e:
         _LOGGER.error("Failed to convert replay response to batch: %s", e)
         raise ValueError(f"Replay data conversion failed: {e}") from e
